add generics to utils
